% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesthresh.R
\name{bayesthresh}
\alias{bayesthresh}
\title{Bayesian wavelet thresholding}
\usage{
bayesthresh(
  x,
  j0 = 3,
  alpha = 0.5,
  beta = 1,
  C1 = NA,
  C2 = NA,
  C1.start = 100,
  family = "Coiflets",
  filter.size = 30,
  wavelet.filter = NULL
)
}
\arguments{
\item{x}{Vector with the signal to be regularized. Its length must be a
power of two.}

\item{j0}{The coarsest resolution level whose detail coefficients are
thresholded. The coefficients of the levels below \code{j0} are left
untouched, as is the usual practice. The default \code{j0 = 3} is the one
of \code{wavethresh::threshold} and the one used by Motta and Montoril
(2026).}

\item{alpha, beta}{The parameters \eqn{\alpha} and \eqn{\beta} of the prior.
They must be non-negative. The defaults \eqn{\alpha = 0.5} and
\eqn{\beta = 1} are the robust choice recommended by Abramovich, Sapatinas
and Silverman (1998).}

\item{C1, C2}{The hyperparameters \eqn{C_1} and \eqn{C_2} of the prior. When
they are left as \code{NA}, the default, they are estimated from the data.
See Details.}

\item{C1.start}{Sets the search interval \eqn{[0, 50\sqrt{C_{1,start}}]} of
the estimation of \eqn{C_1}. It is only used when \code{C1} is \code{NA}.}

\item{family}{The family of wavelets to use, as in \command{\link{wavedec}}.}

\item{filter.size}{The size of the wavelet filter.}

\item{wavelet.filter}{Use this to provide your own filter, together with
\code{family = "Own"}.}
}
\value{
A list with components
\item{fit}{The regularized signal, of the same length as \code{x}.}
\item{coefficients}{The thresholded wavelet coefficients, decomposed down to
  the level zero.}
\item{sigma}{The estimated noise level.}
\item{C1, C2}{The hyperparameters of the prior, supplied or estimated. They
  are \code{NA} when the noise level is degenerate, in which case no
  thresholding is performed.}
}
\description{
Regularizes a signal by the Bayesian wavelet thresholding of
Abramovich, Sapatinas and Silverman (1998), which places a spike and slab
prior on the detail wavelet coefficients and shrinks them towards the
posterior median. It is the regularization step of \command{\link{bwmixreg}}.
}
\details{
The signal is decomposed by the discrete wavelet transform down to the level
zero, and each detail coefficient \eqn{d_{jk}} of the levels
\eqn{j_0, \ldots, J - 1} is assigned the prior
\deqn{\pi_j N(0, v_j^2) + (1 - \pi_j)\delta_0(\theta_{jk}), \qquad
      v_j^2 = 2^{-\alpha j}C_1, \qquad \pi_j = \min(1, 2^{-\beta j}C_2),}
where \eqn{\delta_0} is a point mass at zero. The coefficients are then
replaced by the median of the resulting posterior distribution, which
thresholds them, and the signal is reconstructed.

The noise level is estimated by the median absolute deviation of the detail
coefficients of the finest level. The hyperparameter \eqn{C_1} is estimated
by maximizing the marginal likelihood of the coefficients that survive a
universal hard threshold, and \eqn{C_2} follows from it in closed form.

This function was implemented based on the \code{"BayesThresh"} policy of
\code{wavethresh::threshold.wd}, of G. P. Nason, which is the implementation
used by Motta and Montoril (2026). The estimation of \eqn{C_1} relies on a
one-dimensional minimizer implemented based on \code{Brent_fmin}, the engine
of \command{\link{optimize}}, itself a translation of the algorithm of Brent
(1973). Note that, since the wavelet filters of this package and those of
\pkg{wavethresh} follow different conventions, the coefficients of the two
implementations are indexed differently, and the regularized signals agree
only up to that convention.
}
\examples{
set.seed(123)
n <- 256
t <- (1:n)/n
f <- sin(4*pi*t)
y <- f + rnorm(n, sd = 0.3)

bt <- bayesthresh(y, family = "Coiflets", filter.size = 30)

plot(t, y, col = "grey60", cex = 0.6, xlab = "t", ylab = "y")
lines(t, f, col = 2, lwd = 2, lty = 2)
lines(t, bt$fit, lwd = 2)

c(sigma = bt$sigma, C1 = bt$C1, C2 = bt$C2)

}
\references{
Abramovich, F., Sapatinas, T. and Silverman, B. W. (1998). Wavelet
thresholding via a Bayesian approach. \emph{Journal of the Royal Statistical
Society Series B}, 60(4), 725--749,
\doi{10.1111/1467-9868.00151}.

Brent, R. P. (1973). \emph{Algorithms for Minimization without Derivatives}.
Prentice-Hall, Englewood Cliffs.

Motta, F. C. and Montoril, M. H. (2026). A Bayesian estimation approach for
the wavelet-based mixture regression. \emph{Communications in Statistics -
Simulation and Computation}, 55(6), 2426--2434,
\doi{10.1080/03610918.2025.2470797}.
}
\seealso{
\command{\link{bwmixreg}}, \command{\link{bwregime}},
  \command{\link{wavedec}}, \command{\link{waverec}}
}
\author{
Michel H. Montoril \email{michel@ufscar.br}
}
\keyword{smooth}
