% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwmixreg.R
\name{plot.bwmixreg}
\alias{plot.bwmixreg}
\title{Plot method for the Bayesian wavelet-based mixture regression}
\usage{
\method{plot}{bwmixreg}(
  x,
  band = TRUE,
  data = FALSE,
  estimate = c("median", "mean"),
  main,
  xlab = "Index",
  ylab = "Estimated probability",
  type = "l",
  col = 1,
  band.col = "grey80",
  ...
)
}
\arguments{
\item{x}{A fitted object of class \code{"bwmixreg"}.}

\item{band}{Logical. Should the highest posterior density band be drawn?}

\item{data}{Logical. Should the observed mixture be added to the plot, on a
secondary axis? It is useful to relate the peaks of the weights to the
observations, as in Figure 3 of the paper.}

\item{estimate}{Which point estimate to draw, the posterior \code{"median"}
(the default, as in the paper) or the posterior \code{"mean"}.}

\item{main, xlab, ylab, type, col, band.col}{Graphical parameters with sensible
defaults.}

\item{...}{Further graphical parameters passed to \command{\link{plot}}.}
}
\value{
Nothing. The function is called for its side effect.
}
\description{
Plots the estimated dynamic mixture weights of a
\code{"bwmixreg"} object, together with their highest posterior density
band, as in Figure 3(b) of Motta and Montoril (2026).
}
\details{
The function draws a single panel, so further curves may be superimposed on
it afterwards with \command{\link{lines}} or \command{\link{points}}, which
is how the true weight function is added in the examples of
\command{\link{bwmixreg}}.
}
\examples{
set.seed(321)
n <- 256
t <- (1:n)/n
a <- 0.2*(t < 0.3) + 0.8*(t >= 0.3 & t < 0.7) + 0.3*(t >= 0.7)
z <- rbinom(n, size = 1, prob = a)
y <- z*rnorm(n, mean = 2, sd = 0.5) + (1 - z)*rnorm(n, mean = 0, sd = 0.5)

fit <- bwmixreg(y, nchain = 300, burn = 300, plot = FALSE)

plot(fit)
plot(fit, data = TRUE, band = FALSE)

}
\seealso{
\command{\link{bwmixreg}}
}
\author{
Michel H. Montoril \email{michel@ufscar.br}
}
\keyword{smooth}
