% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwregime.R
\name{plot.bwregime}
\alias{plot.bwregime}
\title{Plot method for the Bayesian regime switching model}
\usage{
\method{plot}{bwregime}(
  x,
  band = TRUE,
  data = FALSE,
  estimate = c("median", "mean"),
  main = "",
  xlab = "Index",
  ylab = "Estimated probability",
  type = "l",
  col = 1,
  band.col = "grey80",
  ...
)
}
\arguments{
\item{x}{A fitted object of class \code{"bwregime"}.}

\item{band}{Logical. Should the highest posterior density band be drawn?}

\item{data}{Logical. Should the observed mixture be added to the plot, on a
secondary axis? It is useful to relate the regimes to the observations, as
in Figures 8 and 9 of the paper.}

\item{estimate}{Which point estimate to draw, the posterior \code{"median"}
(the default, as in the paper) or the posterior \code{"mean"}.}

\item{main, xlab, ylab, type, col, band.col}{Graphical parameters with sensible
defaults.}

\item{...}{Further graphical parameters passed to \command{\link{plot}}.}
}
\value{
Nothing. The function is called for its side effect.
}
\description{
Plots the estimated dynamic mixture weights of a
\code{"bwregime"} object, together with their highest posterior density
band, as in Figures 7 and 9 of Motta and Montoril (2026b).
}
\details{
The function draws a single panel, so further curves may be superimposed on
it afterwards with \command{\link{lines}} or \command{\link{points}}, which
is how the true weight function is added in the examples of
\command{\link{bwregime}}. It is the plot method of
\command{\link{bwmixreg}}, applied to the estimates of this model.
}
\examples{
set.seed(321)
n <- 256
t <- (1:n)/n
a <- 0.05 + 0.9*(t > 0.3 & t < 0.5)
z <- rbinom(n, size = 1, prob = a)
y <- z*rnorm(n, mean = 3, sd = 0.5) + (1 - z)*rnorm(n, mean = 0, sd = 0.5)

fit <- bwregime(y, nchain = 300, burn = 300, lag = 5, plot = FALSE)

plot(fit)
plot(fit, data = TRUE, band = FALSE)

}
\seealso{
\command{\link{bwregime}}, \command{\link{plot.bwmixreg}}
}
\author{
Michel H. Montoril \email{michel@ufscar.br}
}
\keyword{smooth}
