% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmixreg.R
\name{plot.wmixreg}
\alias{plot.wmixreg}
\title{Plot method for wavelet-based mixture regression}
\usage{
\method{plot}{wmixreg}(
  x,
  estimator = NULL,
  regularization = NULL,
  data = FALSE,
  main,
  xlab = "Index",
  ylab = "Estimated probability",
  type = "l",
  ...
)
}
\arguments{
\item{x}{A fitted object of class \code{"wmixreg"}.}

\item{estimator}{Optional subset of the raw estimators to plot.}

\item{regularization}{Optional subset of the regularization methods to plot.}

\item{data}{Logical. Should the transformed responses be added to the plot?
It is only meaningful when a single panel is drawn.}

\item{main, xlab, ylab, type}{Graphical parameters with sensible defaults.}

\item{...}{Further graphical parameters passed to \command{\link{plot}}.}
}
\value{
Nothing. The function is called for its side effect.
}
\description{
Plots the estimated mixing probability curves of a
\code{"wmixreg"} object, arranging the regularization methods along the rows
and the raw estimators along the columns, as in Figure 12 of Montoril,
Pinheiro and Vidakovic (2019).
}
\details{
When more than one panel is drawn, the function sets \code{par(mfrow)} and
restores it on exit. Superimposing curves afterwards with
\command{\link{lines}} or \command{\link{points}} therefore only works when a
single panel is requested, that is, when both \code{estimator} and
\code{regularization} select exactly one element. Otherwise the low-level
calls would be drawn on the restored layout instead of on the last panel.
}
\examples{
set.seed(123)
n <- 400
t <- (1:n)/n
f <- 0.8*t + 0.1
z <- rbinom(n, size = 1, prob = f)
y <- z*rnorm(n, mean = 2) + (1 - z)*rnorm(n, mean = 0)

fit <- wmixreg(y, mu.U = 2, mu.V = 0, J = 4, plot = FALSE)
plot(fit)
plot(fit, estimator = "phi", regularization = "loclin")

}
\seealso{
\command{\link{wmixreg}}
}
\author{
Michel H. Montoril \email{michel@ufscar.br}
}
\keyword{smooth}
