% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.wall.R
\name{predict.cv.wall}
\alias{predict.cv.wall}
\alias{coef.cv.wall}
\title{Predictions and coefficients from a cross-validated wall classifier}
\usage{
\method{predict}{cv.wall}(object, newx, s = c("lambda.min", "lambda.1se"), ...)

\method{coef}{cv.wall}(object, s = c("lambda.min", "lambda.1se"), ...)
}
\arguments{
\item{object}{A fitted object of class \code{"cv.wall"}.}

\item{newx}{Matrix (or data frame) of covariates at which predictions are
to be made. See \command{\link{predict.wall}}.}

\item{s}{Value(s) of the penalty parameter at which predictions are
computed: \code{"lambda.min"} (default, the minimizer of the
cross-validated measure), \code{"lambda.1se"} or a numeric value.}

\item{...}{Further arguments passed to \command{\link{predict.wall}}
(e.g., \code{type}).}
}
\value{
See \command{\link{predict.wall}}.
}
\description{
Obtains predictions and coefficients from a
\command{\link{cv.wall}} object, using the resolution level selected by
the cross-validation and, by default, the value \code{lambda.min} of the
penalty parameter.
}
\examples{
set.seed(1)
n <- 300
x <- matrix(runif(2*n), n, 2)
y <- rbinom(n, 1, 1/(1 + exp(-5*(x[, 1] - 0.5))))
cvfit <- cv.wall(x, y, J = 2:3, filter.size = 8, nfolds = 5)

predict(cvfit, x[1:5, ], type = "response")
predict(cvfit, x[1:5, ], s = "lambda.1se", type = "class")
head(coef(cvfit))

}
\seealso{
\command{\link{cv.wall}}, \command{\link{predict.wall}}
}
\author{
Michel H. Montoril \email{michel@ufscar.br}
}
\keyword{classif}
