% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveBased-package.R
\docType{package}
\name{WaveBased-package}
\alias{WaveBased}
\alias{WaveBased-package}
\title{WaveBased: Wavelet-Based Estimation}
\description{
Provides functions related to the solution of different statistical problems by methods developed under wavelet approach.
}
\author{
\strong{Maintainer}: Michel H. Montoril \email{michel@ufscar.br}

}
\keyword{internal}
